requirements.txt
PySimpleGUI-4-foss